extract common utils for array_slice
